changind dir project and interest
